// src/shared/interface.ts
export enum AssetType {
  BRC20,
  COLLECTIBLE,
  RUNES,
}